Added README. Added hierarchical make. Began implementing Salsa20.
Added a README for placement on github. Implemented a hierarchical
Makefile for separate project and test compilation. Improved an
invalid argument error with a call meh_warn. Began the
implementation of Salsa20 using the files for RC4 (I did this
before the initial import into git--having been using mercurial
previously--which is why it is not on a separate branch.)